Ensure that we use the latest applicable rule in TPascalTZ.Convert.
When rules are sorted by year and then by month, it doesn't guarantee that all applicable rules for the year in question are also in ascending order by month. For example: Rule1 = 2000-2010 Oct; Rule2 = 2005-2010 Mar; for years 2005-2010 both rules are applicable but they are not in ascending order by month (Oct, Mar), so simply using the last rule in the list will not work!