Improved search for applicable rule by finding the most recently activated one out of all rules.
This fixes the problem of not applying rules inherited from the past years (prior to the year of the date in question).

Previous implementation incorrectly assumed that every year starts with DST offset of zero (0) - that is simply not true; also assumed that effect of every rule ends at the end of the year - also not true.