Replaced broken calculation of Days Since AD (DSAD) with Julian Day Number (JDN).
The method of fixing overflowing seconds of day (when it goes negative or beyond number of seconds in 1 day) involved a conversion of the date to DSAD (linear result), incrementing/decrementing DSAD by an appropriate amount of full days to fix overflow of seconds of day, and then converting DSAD back to a date. Functions ElapsedDaysSinceAD and/or ElapsedDaysSinceADToDate had an unknown problem which resulted in certain dates to be encoded/decoded incorrectly. For example, adding 1 day to 1992-12-29 would produce 1992-12-31 (2 days difference).

Calculation of Days Since AD (DSAD) was replaced with Julian Day Number (JDN), implemented according to formulas found on Wikipedia page:
https://en.wikipedia.org/wiki/Julian_day