Removed AOnlyGeoZones parameter from GetTimeZoneNames function.
AOnlyGeoZones parameter would filter the names of time zones based on whether they have '/' character or not. This is not a true filter for geographical time zones, for example it incorrectly classifies zones in 'backward' and 'etcetera' definition files.

There is no clear way of filtering geographical time zones, hence, this feature is not feasible.