Fixed ReloadDatabasePath for Linux OS
On Unix, passing a directory name will result in True with FileExists function (because directory is file too in Unix). So it's valid to check if is really file not directory.
Perhaps it's more beautiful to just put a check on the directory at the beginning, but I'm not sure if this matches the logic of the code. That is it is possible and so, it is possible
    if DirectoryExists(FDatabasePath) then
      ParseDatabaseFromDirectory(FDatabasePath) 
    else if FileExists(FDatabasePath) then
      ParseDatabaseFromFile(FDatabasePath)